Apply draft changes to components
