--- packages/react/src/components/DefaultEmptyFieldEditingComponentsOptimized.tsx
@@ -1,55 +1,0 @@
-import React from 'react';
-
-interface EmptyFieldComponentProps {
-  [key: string]: unknown;
-  tag?: React.ElementType;
-  className?: string;
-}
-
-/**
- * The DefaultEmptyFieldEditingComponentText component - Optimized version using JSX instead of React.createElement.
- *
- * This is a modernized version that uses JSX syntax for better readability and
- * maintainability, replacing React.createElement calls with native JSX.
- *
- * @param {object} props - The props for the component.
- * @public
- */
-export const DefaultEmptyFieldEditingComponentTextOptimized = (props: EmptyFieldComponentProps) => {
-  const Tag = (props.tag || 'span') as React.ElementType;
-  return (
-    <Tag {...props} suppressHydrationWarning={true}>
-      [No text in field]
-    </Tag>
-  );
-};
-
-/**
- * The DefaultEmptyFieldEditingComponentImage component (unchanged - already uses JSX).
- * Re-exported here for completeness in the Optimized module.
- *
- * @param {object} props - The props for the component.
- * @public
- */
-export const DefaultEmptyFieldEditingComponentImageOptimized = (
-  props: EmptyFieldComponentProps
-) => {
-  const inlineStyles = {
-    minWidth: '48px',
-    minHeight: '48px',
-    maxWidth: '400px',
-    maxHeight: '400px',
-    cursor: 'pointer',
-  };
-
-  return (
-    <img
-      {...props}
-      suppressContentEditableWarning={true}
-      alt=""
-      src='data:image/svg+xml,%3Csvg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" viewBox="0 0 240 240" style="enable-background:new 0 0 240 240;" xml:space="preserve"%3E%3Cstyle type="text/css"%3E .st0%7Bfill:none;%7D .st1%7Bfill:%23969696;%7D .st2%7Bfill:%23FFFFFF;%7D .st3%7Bfill:%23FFFFFF;stroke:%23FFFFFF;stroke-width:0.75;stroke-miterlimit:10;%7D%0A%3C/style%3E%3Cg%3E%3Crect class="st0" width="240" height="240"/%3E%3Cg%3E%3Cg%3E%3Crect x="20" y="20" class="st1" width="200" height="200"/%3E%3C/g%3E%3Cg%3E%3Ccircle class="st2" cx="174" cy="67" r="14"/%3E%3Cpath class="st2" d="M174,54c7.17,0,13,5.83,13,13s-5.83,13-13,13s-13-5.83-13-13S166.83,54,174,54 M174,52 c-8.28,0-15,6.72-15,15s6.72,15,15,15s15-6.72,15-15S182.28,52,174,52L174,52z"/%3E%3C/g%3E%3Cpolyline class="st3" points="29.5,179.25 81.32,122.25 95.41,137.75 137.23,91.75 209.5,179.75 "/%3E%3C/g%3E%3C/g%3E%3C/svg%3E'
-      className={`scEmptyImage ${props.className || ''}`}
-      style={inlineStyles}
-    />
-  );
-};
