+++ packages/react/src/enhancers/withClientPlaceholder.tsx
@@ -1,0 +1,7 @@
+'use client';
+import { withAppPlaceholder } from './withAppPlaceholder';
+
+/**
+ * withAppPlaceholder import for client-side context and client components
+ */
+export const withClientPlaceholder = withAppPlaceholder;
